requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
Babel==2.5.3
billiard==3.5.0.4
braintree==3.45.0
celery==4.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-datepicker-widget==0.1
django-bootstrap3==10.0.1
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap4==0.0.6
django-braces==1.13.0
django-bulk-update==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-fontawesome==1.0
django-form-utils==1.0.3
django-libs==1.67.13
django-phonenumber-field==2.0.0
django-tables2==1.21.2
dominate==2.3.1